Remade smach_nav_channel.py reduced code, that uses smach_gate.py. Also contains backup of original smach_nav_channel.py
